requirements.txt
# requirements.txt
# Save this file in: lost_found_reunion/requirements.txt

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Text Embeddings (Sentence-Transformers)
sentence-transformers==2.2.2
torch==2.1.1
torchvision==0.16.1

# Image Embeddings (CLIP)
transformers==4.35.2
open-clip-torch==2.23.0

# Vector Database
chromadb==0.4.18

# Local LLM
ollama==0.1.6

# Web UI
streamlit==1.28.0

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
scikit-learn==1.3.2